refactor: bikeshare.py - Change get_filters() to work with get_user_input
get_filters() has the task to get from the user three values that
filter the data according to user input. For each of these values the
same raw input was repeated making up for redundant code. This has
been changed to three calls to get_user_input, making the code
shorter, easier to read and understand.